[Refactor] Fix BE UT on MacOS (#71310)

+++ be/test/exec/pipeline/mem_limited_chunk_queue_test.cpp
@@ -86,7 +86,7 @@ public:
         chunk->append_column(std::move(col), 0);
         return chunk;
     }
-    size_t _next_value = 0;
+    uint64_t _next_value = 0;
     size_t _chunk_size;
 };
